Report when the first error was found.
It is usually useful to know how many iterations was used before an error was
found instead of how many iterations was used altogether. That say something
about how easy the error was to find. We also report how many additional errors
we found in the shrinking process as that says something about how common this
error is.